Fix unit test execution for Python 3.2 ( closes #83 )
The pygame.tests.testutils.async_sub module sets the stdout pipe to O_NONBLOCK.
Yet for linux a pipe set as O_NONBLOCK returns immediately from a select call,
even with nothing to read. This mucks up the Python 3.2 pipe read, which
expects something, and fails (a bug) when there is nothing.

Setting a socket to O_NONBLOCK is a good idea. A remote connection may
have a transmission error, which may trigger the select statement to
return, but subsequently invalidate the data so 0 bytes are returned on a read.
But a pipe, in this case, is inter-task; there should be no transmission errors.
Therefore the O_NONBLOCK should be unnecessary, and safely removed.