Remove cross library FILE * dependencies in imageext.c
In Windows there is, unfortunately, multiple versions of the
C runtime library. None of them are compatible with the others.
This changeset removes the need for the jpeg and png DLLs
to use the same C library as does the imageext module, by
keeping file IO within imageext. It is hoped this is the last
change needed to allow Pygame's Windows prebuilt libraries to link
to a different C runtime library, preferably msvcrt.dll, than
that used by the Python interpreter and Pygame extension modules.